+++ src/keybinds/config.rs
@@ -137,6 +137,8 @@ pub struct LogTabKeybindsConfig {
     pub push_new: Option<Keybind>,
     pub push_all: Option<Keybind>,
     pub push_all_new: Option<Keybind>,
+    pub push_change: Option<Keybind>,
+    pub push_named: Option<Keybind>,
     pub fetch: Option<Keybind>,
     pub fetch_all: Option<Keybind>,
 }